% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.dataset.R
\name{filter.dataset}
\alias{filter.dataset}
\title{Basic quality control of variants.}
\usage{
filter.dataset(
  input.file = input,
  plink.files = FALSE,
  bed.file = bed,
  bim.file = bim,
  fam.file = fam,
  remove.related = TRUE,
  kinship.coefficient = value,
  geno.value = geno,
  maf.value = maf,
  mind.value = mind,
  limit.LD = FALSE,
  indep.pairwise.kb = window.size,
  indep.pairwise.ct = step.size,
  indep.pairwise.r2 = r2.threshold,
  plink.exec = plink.path,
  plink2.exec = plink2.path
)
}
\arguments{
\item{input.file}{is either a VCF or BCF file present in the working directory. File path is also accepted, but not suggested.}

\item{plink.files}{(required) asks whether the input files are in PLINK format. Default is FALSE.}

\item{bed.file}{should be provided if plink.files = TRUE}

\item{bim.file}{should be provided if plink.files = TRUE}

\item{fam.file}{should be provided if plink.files = TRUE}

\item{remove.related}{(required) asks whether kinship coefficients will be calculated and related samples will be removed.}

\item{kinship.coefficient}{should be provided if remove.related = TRUE.  Based on kingrelatedness.com/manual, a kinship coefficient range of >0.354 corresponds to duplicate/MZ twins, [0.177, 0.354] to 1st-degree relationships, [0.0844, 0.177] to 2nd-degree relationships, and [0.0442, 0.0884] to 3rd-degree relationships.}

\item{geno.value}{is the maximum threshold of missingness per variant.}

\item{maf.value}{is the maximum threshold of missingness per variant.}

\item{mind.value}{is the percentage threshold of individuals with missing data}

\item{limit.LD}{asks whether the function should prune the markers that are in linkage equilibrium.}

\item{indep.pairwise.kb}{should be set if limit.LD is TRUE. This is the window size in variant count or kilobase units.}

\item{indep.pairwise.ct}{should be set if limit.LD is TRUE. This is the variant count to shift the window at each end of step.}

\item{indep.pairwise.r2}{should be set if limit.LD is TRUE. This is the pairwise threshold. At each step, pairs of variants in the current window with squared correlation greater than the threshold are noted and the variants are pruned from the window until no such pairs remain.}
}
\description{
This function performs QC of the variant dataset based on commonly used filters.
}
\examples{
filter.dataset("myvcf.vcf", plink.files = FALSE, remove.related = TRUE, kinship.coefficient = 0.0844, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = FALSE)
filter.dataset("myvcf.vcf", plink.files = FALSE, remove.related = FALSE, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = TRUE, indep.pairwise.kb = 200, indep.pairwise.ct = 30, indep.pairwise.r2 = 0.5)
filter.dataset(plink.files = TRUE, bed.file = "mybed.bed", bim.file = "mybim.bim", fam.file = "myfam.fam", remove.related = FALSE, geno.value = 0.5, maf.value = 0.05, mind.value = 0.1, limit.LD = TRUE, indep.pairwise.kb = 300, indep.pairwise.ct = 25, indep.pairwise.r2 = 0.4)
}
