% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractmarkers.R
\name{extract_markers}
\alias{extract_markers}
\title{Extracts a set of markers from PLINK, VCF, or BCF files}
\usage{
extract_markers(
  input.file,
  snps.list = NULL,
  pos.list = NULL,
  bed.file = NULL,
  bim.file = NULL,
  fam.file = NULL,
  output.dir = "marker_outputs",
  merged.file = "final_merged.vcf"
)
}
\arguments{
\item{input.file}{(character) The file path to the genetic dataset. Accepts vcf or bcf files.}

\item{snps.list}{(character) The file path to the text file containing the list of markers (rsID) to be extracted. Each line will be read as one marker. Should be a text file.}

\item{pos.list}{(character) The file path to the text file containing the list of markers to be extracted. If using rsIDs, each line should correspond to one marker. If using POS (positions), there should be 4 columns: [1] chromosome numner (integer), [2] starting base position, [3] final base position, and [4] output file name.}

\item{bed.file}{(character) The file path for the bed file. Should be provided if plink.files = TRUE}

\item{bim.file}{(character) The file path for the bim file. Should be provided if plink.files = TRUE}

\item{fam.file}{(character) The file path for the fam file. Should be provided if plink.files = TRUE}
}
\description{
The 'extract_markers' function extracts a set of markers, either using their rsID or chromosome position, from PLINK, VCF, or BCF files. 
In using rsIDs for extraction, a text file containing one marker per line is accepted. When using GRCh38/GRCh37 positions, the text file should contain 4 columns:
the first column should contain the chromosome number only (e.g. "1" and not "chr1"); the second column should contain the starting bp position; 
the third column the final bp position; and the fourth column should specify the desired output name with no file extension (e.g. "snpsfromchr1").
}
\examples{
extractmarkers(input.file = "test.vcf", snps.list = "markers.txt")
extractmarkers(bed.file = "data.bed", bim.file = "data.bim", fam.file = "data.fam", pos.list = "markers.txt")
}
