% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractmarkers.R
\name{extractmarkers}
\alias{extractmarkers}
\title{Extracts a set of markers from PLINK, VCF, or BCF files}
\usage{
extractmarkers(
  input.file = input,
  plink.files = FALSE,
  bed.file = bed,
  bim.file = bim,
  fam.file = fam,
  snps = snps.file,
  output = output.name,
  plink.exec = plink.path
)
}
\arguments{
\item{input.file}{is either a VCF or BCF file present in the working directory. File path is also accepted, but not suggested.}

\item{plink.files}{(required) asks whether the input files are in PLINK format. Default is FALSE.}

\item{bed.file}{should be provided if plink.files = TRUE}

\item{bim.file}{should be provided if plink.files = TRUE}

\item{fam.file}{should be provided if plink.files = TRUE}

\item{snps}{(required) is a text file containing the list of markers to be extracted. Each line should correspond to one marker.}

\item{output}{(required) is the user's desired output name.}

\item{plink.exec}{(required) takes in the file path for the plink executable. Suggestion is to have plink.exe in the working directory.}
}
\description{
Extracts a set of markers from PLINK, VCF, or BCF files
}
\examples{
extractmarkers(input.file = "mygenotype.vcf", plink.files = FALSE, snps = "listsnps.txt", output = "success", plink.exec = "./plink.exe")
extractmarkers(plink.files = TRUE, bed.file = "data.bed", bim.file = "data.bim", fam.file = "data.fam", snps = "listsnps.txt", output = "success", plink.exec = "./softwares/plink.exe")
}
